refactor time.test.ts into smaller files
